Cherry-pick 283286.365@safari-7620-branch (e7b7957de026). https://bugs.webkit.org/show_bug.cgi?id=281912
DocumentFontLoader::fontLoadingTimerFired() must keep a Ref of Document on the stack
https://bugs.webkit.org/show_bug.cgi?id=281912
rdar://138215892

Reviewed by Chris Dumez.

DocumentFontLoader::fontLoadingTimerFired() calls CachedResourceLoader::loadDone(),
which holds a RefPtr to the Document. It seems that in certain cases (like the
reproduction case in the radar), this is the only Ref keeping the Document alive.
So when the function ends, the Document is destroyed. Then, when fontLoadingTimerFired()
calls Document::frame() with it's WeakRef m_document, there is a crash since the
Document has been destroyed.

Since Document owns DocumentFontLoader, we make DocumentFontLoader forward its refcounting
to its owning Document. Then we ensure that the Document is alive by holding a RefPtr to
the DocumentFontLoader itself at the beginning of fontLoadingTimerFired().

* LayoutTests/http/tests/security/document-cached-font-loading-expected.txt: Added.
* LayoutTests/http/tests/security/document-cached-font-loading.html: Added.
* LayoutTests/http/tests/security/resources/document-cached-font-loading-helper.html: Added.
* LayoutTests/http/tests/security/resources/font.ttf: Added.
This is the test created based on the reproduction case in the Radar.

* Source/WebCore/dom/DocumentFontLoader.cpp:
(WebCore::DocumentFontLoader::fontLoadingTimerFired):

Canonical link: https://commits.webkit.org/282416.346@webkitglib/2.46