[GStreamer] Account for internal pauses because of buffering https://bugs.webkit.org/show_bug.cgi?id=287881
Reviewed by Xabier Rodriguez-Calvar.

The original patch[1] needs a follow-up to ensure that the simulated
paused state because of buffering isn't reported to the upper layers as
a formal pause (requested by the user).

Author: suresh-khurdiya-infosys <skhurdiya.ext@libertyglobal.com>

[1] https://commits.webkit.org/290606@main

* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::paused const): Early bail out (returning false, meaning "not paused") also when we're simulating a pause internally because of waiting for buffering.

Canonical link: https://commits.webkit.org/290782@main